pyproject.toml
[project]
name = "website-change-discord"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fake-useragent>=2.2.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "schedule>=1.2.2",
    "watchdog>=4.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[project.scripts]
add_site = "src.add_site:main"
monitor = "src.main:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = ["."]
